+++ src/services/colors/index.js
@@ -1,2 +1,2 @@
-export { requiresLightText } from './color_utils';
+export { isColorDark } from './is_color_dark';
 export { VISUALIZATION_COLORS } from './visualization_colors';
