+++ src/services/colors/is_color_dark.js
@@ -1,17 +1,18 @@
 /**
- * This function calculates if the text on a given background color needs to
- * be light or dark for better readability.
+ * This function calculates if the specified color is "dark", which usually means
+ * you need light text if you use it as a background color to fulfill WCAG contrast
+ * requirement.
  * The color must be specified via its red, green and blue value in the range of
  * 0 to 255.
- * The formular is based on this Stackoverflow answer: https://stackoverflow.com/a/3943023
+ * The formula is based on this Stackoverflow answer: https://stackoverflow.com/a/3943023
  * which itself is based upon the WCAG recommendation for color contrast.
  *
  * @param {number} red The red component in the range 0 to 255
  * @param {number} green The green component in the range 0 to 255
  * @param {number} blue The blue component in the range 0 to 255
- * @returns {boolean} True if the background requires light text on it, false otherwise.
+ * @returns {boolean} True if the color is dark, false otherwise.
  */
-function requiresLightText(red, green, blue) {
+function isColorDark(red, green, blue) {
   const [r, g, b] = [red, green, blue]
     .map(c => c / 255.0)
     .map(c => c <= 0.03928 ? c / 12.92 : ((c + 0.055) / 1.055) ** 2.4);
@@ -19,4 +20,4 @@ function requiresLightText(red, green, blue) {
   return luminance <= 0.179;
 }
 
-export { requiresLightText };
+export { isColorDark };
